+++ Eby_GlobalConstants.py
